+++ components/Nav.js
@@ -18,7 +18,7 @@ const pages = [
 ]
 
 const Nav = () => {
-  const { user, signout } = useAuth()
+  const { data, signout } = useAuth()
   const { width } = useWindowSize()
 
   const [navbarIsOpen, setNavbarIsOpen] = useState(false)
@@ -50,7 +50,7 @@ const Nav = () => {
                 </Link>
               ))}
 
-              {!user && (
+              {!data && (
                 <>
                   <Link passHref href='/login'>
                     <Button appearance='subtle'>
@@ -65,7 +65,7 @@ const Nav = () => {
                 </>
               )}
 
-              {user && (
+              {data && (
                 <Popup
                   isOpen={popupIsOpen}
                   onClose={() => setPopupIsOpen(false)}
@@ -74,7 +74,7 @@ const Nav = () => {
                   content={() => (
                     <div style={{ width: '15rem' }}>
                       <MenuGroup>
-                        <Link passHref href='/profile'>
+                        <Link passHref href='/u'>
                           <ButtonItem>
                             Profile
                           </ButtonItem>
@@ -90,7 +90,7 @@ const Nav = () => {
                       {...triggerProps}
                     >
                       <Avatar
-                        src={user.photoURL}
+                        src={data.picture}
                         onClick={() => setPopupIsOpen(!popupIsOpen)}
                       />
                     </div>
@@ -107,7 +107,7 @@ const Nav = () => {
         <MenuGroup style={{ backgroundColor: '#f6f6f6' }}>
           <Section title='Pages'>
             {pages.map((page, i) => (
-              <Link passHref key={i} href={page.href} passHref>
+              <Link passHref key={i} href={page.href}>
                 <ButtonItem isDisabled={!page.href}>
                   {page.label}
                 </ButtonItem>
@@ -116,10 +116,10 @@ const Nav = () => {
           </Section>
 
           <Section title='Account' hasSeparator>
-            {user ? (
+            {data ? (
               <>
-                <Link passHref href='/profile'>
-                  <ButtonItem elemAfter={<Avatar src={user.photoURL} />}>
+                <Link passHref href='/u'>
+                  <ButtonItem elemAfter={<Avatar src={data.photoURL} />}>
                     My Profile
                   </ButtonItem>
                 </Link>