+++ modules/firebase.js
@@ -3,6 +3,7 @@ import 'firebase/database'
 import 'firebase/auth'
 import 'firebase/analytics'
 import 'firebase/firestore'
+import 'firebase/functions'
 
 if (!firebase.apps.length) {
   firebase.initializeApp({
@@ -19,6 +20,15 @@ if (!firebase.apps.length) {
   // If on client-side
   if (typeof window !== 'undefined') {
     firebase.analytics()
+
+    if (window.location.hostname === 'localhost') {
+      firebase.firestore().settings({
+        host: 'localhost:8080',
+        ssl: false
+      })
+      firebase.functions().useFunctionsEmulator('http://localhost:5001')
+    }
   }
 }
+
 export default firebase