+++ pages/404.js
@@ -1,13 +1,18 @@
-import styled from 'styled-components'
 import Layout from '../components/Layout'
-import Button from '../components/Button'
+import { fontSize } from '@atlaskit/theme'
+import Link from 'next/link'
+import Button from '@atlaskit/button'
 
 const Home = () => (
   <Layout title='404'>
     <h1>Page not found (404)</h1>
-    <h2>
-      We couldn't find that page, <a href='/missions'>go back to your missions</a>?
-    </h2>
+    <p style={{ marginBottom: '1rem', fontSize: fontSize() * 1.2 }}>
+      <span>Couldn't find the requested page. </span>
+    </p>
+
+    <Link href='/missions' passHref>
+      <Button appearance='primary'>Go to your missions</Button>
+    </Link>
   </Layout>
 )