+++ pages/create-account.js
@@ -1,16 +1,46 @@
 import Layout from '../components/Layout'
 import { useAuth } from '../modules/useAuth'
-import Form, { Field, FormFooter, HelperMessage } from '@atlaskit/form'
+import Form, { Field, FormFooter, HelperMessage, ErrorMessage, FormSection } from '@atlaskit/form'
 import Button, { ButtonGroup } from '@atlaskit/button'
 import TextField from '@atlaskit/textfield'
 import firebase from '../modules/firebase'
 import { useRouter } from 'next/router'
+import SectionMessage from '@atlaskit/section-message'
+import { useEffect } from 'react'
+import styled from 'styled-components'
 
 const CreateAccount = () => {
   const { user, data } = useAuth()
   const router = useRouter()
 
-  if (!user) {
+  // Redirect if there is already a username
+  useEffect(() => {
+    if (data) {
+      router.push('/u')
+    }
+  }, [data])
+
+  // Redirect if not signed in
+  useEffect(() => {
+    if (user === false) {
+      router.push('/signup')
+    }
+  }, [user])
+
+  async function createAccount (data) {
+    return firebase.functions()
+      .httpsCallable('setUsername')({
+        username: data.username
+      })
+      .then(() => {
+        router.push('/u')
+      }).catch(err => {
+        console.log({ message: err.message })
+        return { username: err.message }
+      })
+  }
+
+  if (user === null) {
     return (
       <Layout title='Create Account' loading />
     )
@@ -18,32 +48,28 @@ const CreateAccount = () => {
 
   return (
     <Layout title='Create Account'>
-      <p>
-        We found some information from <em>{user.providerData[0].providerId}</em>
-      </p>
-
-      <Form onSubmit={data => {
-        return firebase.firestore()
-          .doc(`/users/${user.uid}`)
-          .set(data)
-          .then(() => { router.push('/missions') })
-      }}
-      >
+      <Form onSubmit={createAccount}>
         {({ formProps, submitting }) => (
           <form {...formProps}>
-            <Field name='username' defaultValue={user.displayName} label='Display name' isRequired>
-              {({ fieldProps }) => <TextField {...fieldProps} />}
-            </Field>
-            <HelperMessage>
-              This will be the way your account is shown to other users, for example, next to your public mission solutions.
-            </HelperMessage>
+            <FormSection title='Choose a username'>
+              <Gap />
+              <SectionMessage appearance='warning'>
+                <p>You won't be able to change your username later.</p>
+              </SectionMessage>
+              <Gap />
 
-            <Field name='fullname' label='Full name'>
-              {({ fieldProps }) => <TextField {...fieldProps} />}
-            </Field>
-            <HelperMessage>
-              We'll put this on your profile page.
-            </HelperMessage>
+              <Field name='username' label='Username' isRequired>
+                {({ fieldProps, error }) => (
+                  <>
+                    <TextField {...fieldProps} />
+                    {error && <ErrorMessage>{error}</ErrorMessage>}
+                  </>
+                )}
+              </Field>
+              <HelperMessage>
+                <p>This will be the way your account is shown to other users, for example, next to your public mission solutions.</p>
+              </HelperMessage>
+            </FormSection>
 
             <FormFooter>
               <ButtonGroup>
@@ -58,5 +84,9 @@ const CreateAccount = () => {
     </Layout>
   )
 }
+
+const Gap = styled.div`
+  padding: .5rem;
+`
 
 export default CreateAccount