+++ pages/missions.js
@@ -8,13 +8,8 @@ import { useRouter } from 'next/router'
 import { useEntities } from '../modules/useEntities'
 
 const Missions = () => {
-  const { user, data } = useAuth()
+  const { user } = useAuth()
   const router = useRouter()
-
-  if (data === false) {
-    router.push('/create-account')
-  }
-
   const after = router.query.after
   const missions = useEntities('mission', 'missions', after)