+++ pages/signup.js
@@ -1,33 +1,28 @@
 import Layout from '../components/Layout'
 import firebase from '../modules/firebase'
 import StyledFirebaseAuth from 'react-firebaseui/StyledFirebaseAuth'
-import { useAuthState } from 'react-firebase-hooks/auth'
 import styled from 'styled-components'
 import Router from 'next/router'
 import { useEffect } from 'react'
+import { useAuth } from '../modules/useAuth'
 
 const uiConfig = {
   signInFlow: 'redirect',
-  signInSuccessUrl: '/missions',
+  signInSuccessUrl: '/create-account',
   signInOptions: [
-    {
-      provider: firebase.auth.GithubAuthProvider.PROVIDER_ID,
-      scopes: [
-        'user'
-      ]
-    }
+    firebase.auth.GithubAuthProvider.PROVIDER_ID
   ]
 }
 
 const SignUp = () => {
   // TODO: handle error
-  const [user, initialising, error] = useAuthState(firebase.auth())
+  const { user } = useAuth()
 
   useEffect(() => {
-    if (user && !initialising) {
-      Router.push('/profile?first=1')
+    if (user) {
+      Router.push('/u')
     }
-  }, [user, initialising])
+  }, [user])
 
   return (
     <Layout title='Sign Up'>