+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2020-04-26
+*This update was a real headache so please be nice to me for the next couple of days*
+### Added 
+- `/u` will redirect to `/u/[signed-in-username]`
+- Firebase emulators for dev experience
+- Firebase functions `setUsername` ensures that users won't set their name to be the same as others
+### Changed
+- Moved profile page to `/u/[username]`
+- All data about users should now be saved in firestore rather than firebase auth
+## Removed
+- `/profile` page (same as `/u/[my username]`)
+
 ## [1.1.3] - 2020-04-24
 - Working on pagination but its not done yet
 ### Added